backend/requirements.txt
django>=5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
djangorestframework-camel-case>=1.4
django-cors-headers>=4.3
django-allauth>=0.60
dj-rest-auth>=5.0
django-ratelimit>=4.1
django-filter>=24.1

celery>=5.3
redis>=5.0

requests>=2.31
beautifulsoup4>=4.12
lxml>=5.1
python-magic-bin>=0.4
urllib3>=2.1

scikit-learn>=1.4
joblib>=1.3
numpy>=1.26
pandas>=2.2

reportlab>=4.1
Pillow>=10.2

openai>=1.12

python-dotenv>=1.0
bleach>=6.1
PyJWT>=2.8
pyotp>=2.9
qrcode>=7.4

gunicorn>=21.2
whitenoise>=6.6

pytest>=8.0
pytest-django>=4.8
factory-boy>=3.3
